+++ tilegame/src/components/animation.hpp
@@ -8,7 +8,7 @@ namespace tilegame::components
 {
     struct Animation
     {
-        double clock;
+        float clock;
         std::size_t current_frame_idx;
         std::vector<engine::sprite::SpriteFrame> frames;
