+++ tilegame/src/components/timer.hpp
@@ -8,21 +8,21 @@ namespace tilegame::components
 {
     struct Timer
     {
-        double time_total;
-        double time_left;
+        float time_total;
+        float time_left;
         bool repeat;
 
         Timer() = default;
-        Timer(double time_total, bool repeat) : time_total(time_total), time_left(time_total), repeat(repeat) {}
+        Timer(float time_total, bool repeat) : time_total(time_total), time_left(time_total), repeat(repeat) {}
     };
 
     struct TimerEventArgs
     {
         inline static const std::string EVENT_TYPE = "timer";
-        double duration;
+        float duration;
         bool repeated;
 
         TimerEventArgs() {}
-        TimerEventArgs(double duration, bool repeated) : duration(duration), repeated(repeated) {}
+        TimerEventArgs(float duration, bool repeated) : duration(duration), repeated(repeated) {}
     };
 } // namespace tilegame
